Simplify action widget api
This changes the action widget take a list of items and a list of action bar actions. This simplifies using the action widget service and lets us move code action specific code (such as documentation items) up into the codeActionUi
